+++ libsolidity/codegen/ABIFunctions.h
@@ -134,6 +134,15 @@ private:
 	/// otherwise an assertion failure.
 	std::string cleanupFunction(Type const& _type, bool _revertOnFailure = false);
 
+	/// Performs cleanup after reading from a potentially compressed storage slot.
+	/// The function does not perform any validation, it just masks or sign-extends
+	/// higher order bytes or left-aligns (in case of bytesNN).
+	/// The storage cleanup expects the value to be right-aligned with potentially
+	/// dirty higher order bytes.
+	/// @param _splitFunctionTypes if false, returns the address and function signature in a
+	/// single variable.
+	std::string cleanupFromStorageFunction(Type const& _type, bool _splitFunctionTypes);
+
 	/// @returns the name of the function that converts a value of type @a _from
 	/// to a value of type @a _to. The resulting vale is guaranteed to be in range
 	/// (i.e. "clean"). Asserts on failure.
@@ -232,6 +241,19 @@ private:
 	std::string abiDecodingFunctionStruct(StructType const& _type, bool _fromMemory);
 	/// Part of @a abiDecodingFunction for array types.
 	std::string abiDecodingFunctionFunctionType(FunctionType const& _type, bool _fromMemory, bool _forUseOnStack);
+
+	/// @returns a function that reads a value type from storage.
+	/// Performs bit mask/sign extend cleanup and appropriate left / right shift, but not validation.
+	/// @param _splitFunctionTypes if false, returns the address and function signature in a
+	/// single variable.
+	std::string readFromStorage(Type const& _type, size_t _offset, bool _splitFunctionTypes);
+
+	/// @returns a function that extracts a value type from storage slot that has been
+	/// retrieved already.
+	/// Performs bit mask/sign extend cleanup and appropriate left / right shift, but not validation.
+	/// @param _splitFunctionTypes if false, returns the address and function signature in a
+	/// single variable.
+	std::string extractFromStorageValue(Type const& _type, size_t _offset, bool _splitFunctionTypes);
 
 	/// @returns the name of a function used during encoding that stores the length
 	/// if the array is dynamically sized (and the options do not request in-place encoding).